Move planner::Sql -> sql::Sql
