+++ src/lib/engine.rs
@@ -2,6 +2,7 @@ use std::str::FromStr;
 
 use crate::lang::{Lang, LangType};
 use crate::planner;
+use crate::sql::Sql;
 use crate::value::PqlValue;
 
 pub fn evaluate(sql: &str, input: &str, from: &str, to: &str) -> anyhow::Result<String> {
@@ -9,7 +10,7 @@ pub fn evaluate(sql: &str, input: &str, from: &str, to: &str) -> anyhow::Result<
     let to_lang_type = LangType::from_str(&to)?;
     let mut lang = Lang::from_as(&input, from_lang_type)?;
 
-    let sql = planner::Sql::from_str(&sql)?;
+    let sql = Sql::from_str(&sql)?;
 
     let result = planner::evaluate(sql, lang.data);
     lang.to = to_lang_type;
@@ -36,7 +37,7 @@ pub fn dumps(data: PqlValue, to: &str) -> anyhow::Result<String> {
 }
 
 pub fn query_evaluate(data: PqlValue, sql: &str) -> anyhow::Result<PqlValue> {
-    let sql = planner::Sql::from_str(&sql)?;
+    let sql = Sql::from_str(&sql)?;
     let data = PqlValue::from(data);
     let value = planner::evaluate(sql, data);
     Ok(value)
