+++ src/lib/sql/sql.rs
@@ -1,0 +1,25 @@
+use std::str::FromStr;
+
+use crate::parser;
+pub use crate::sql::clause::Limit;
+pub use crate::sql::clause::OrderBy;
+use crate::sql::Field;
+pub use crate::sql::WhereCond;
+
+#[derive(Debug, Default, Clone, PartialEq)]
+pub struct Sql {
+    pub select_clause: Vec<Field>,
+    pub from_clause: Vec<Field>,
+    pub left_join_clause: Vec<Field>,
+    pub where_clause: Option<Box<WhereCond>>,
+    pub orderby: Option<OrderBy>,
+    pub limit: Option<Limit>,
+}
+
+impl FromStr for Sql {
+    type Err = anyhow::Error;
+
+    fn from_str(s: &str) -> anyhow::Result<Self> {
+        parser::select_statement::from_str(s)
+    }
+}
